# util: ignore some sql in statement summary