<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Problem Summary:

Ignore some sql in statement summary tables: begin, use, explain, set, and other.
Because these statements won't cause performance problems, but they will make some noises.

### What is changed and how it works?

What's Changed:
Ignore some sql in statement summary tables: begin, use, explain, set, and other.

How it Works:

### Related changes

- PR to update `pingcap/docs`/`pingcap/docs-cn`:
- Need to cherry-pick to the release branch

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Manual test (add detailed scripts or steps below)
```
mysql> use test;
No connection. Trying to reconnect...
Connection id:    1
Current database: *** NONE ***

Database changed
mysql> use test;
Database changed
mysql> explain select 1;
+-------------------+---------+------+---------------+---------------+
| id                | estRows | task | access object | operator info |
+-------------------+---------+------+---------------+---------------+
| Projection_3      | 1.00    | root |               | 1->Column#1   |
| └─TableDual_4     | 1.00    | root |               | rows:1        |
+-------------------+---------+------+---------------+---------------+
2 rows in set (0.00 sec)

mysql> explain analyze select 1;
+-------------------+---------+---------+------+---------------+-----------------------------------------+---------------+---------+------+
| id                | estRows | actRows | task | access object | execution info                          | operator info | memory  | disk |
+-------------------+---------+---------+------+---------------+-----------------------------------------+---------------+---------+------+
| Projection_3      | 1.00    | 1       | root |               | time:3.158µs, loops:2, Concurrency:OFF  | 1->Column#1   | 0 Bytes | N/A  |
| └─TableDual_4     | 1.00    | 1       | root |               | time:690ns, loops:2                     | rows:1        | N/A     | N/A  |
+-------------------+---------+---------+------+---------------+-----------------------------------------+---------------+---------+------+
2 rows in set (0.00 sec)

mysql> select stmt_type, digest_text from performance_schema.events_statements_summary_by_digest\G
*************************** 1. row ***************************
  stmt_type: explain
digest_text: explain analyze select ?
*************************** 2. row ***************************
  stmt_type: select
digest_text: select database ( )
*************************** 3. row ***************************
  stmt_type: show
digest_text: show tables
*************************** 4. row ***************************
  stmt_type: show
digest_text: show databases
*************************** 5. row ***************************
  stmt_type: select
digest_text: select variable_value from mysql . tidb where variable_name = ?
5 rows in set (0.00 sec)
```

Side effects

N/A

### Release note <!-- bugfixes or new feature need a release note -->

- Ignore SET, USE, EXPLAIN, BEGIN statements in statement summary.