@AilinKid 2020-04-02
> I don't think this change is a good idea. At a minimum it is not MySQL compatible behavior (these statements are in statement summary).
> 
> Users rely on this summary table to understand a complete profile of server activity. I would want to know if my applications are sending noisy statements to the server, because even if they are very low cost in TiDB, the network protocol is synchronous. It could indicate that I need to tune a driver to be less noisy (JDBC for example is notoriously bad by default). The latency will add up for applications.
> 
> **Edit:** The "mysql" way to ignore these is sys schema (sys.statement_analysis is ordered by total execution time).

PTAL @djshow832