@djshow832 2020-04-02
I discussed with the person who suggested it, and we decided not to change it. So I'll close it now.